Add explicit/implicit grab demo
This demo shows a bug in mutter.

That is, when a nested popup window is created with explicit grab,
it can be hidden by destroying xdg_popup and xdg_surface objects
and committing null buffer to the wl_surface (wl_surface is kept).

But once it is created without explicit grab, it is not hidden

To use this demo, compile it with:
  $ make demo

Then run with explicit grab
  $ ./demo --with-grab

Once demo starts, a light-blue square appears. Then,
1) Press with a left mouse button on the square and release
2) Another yellow square will appear (it will be created as a popup window)
3) Press again on the appeared square or on the existing one.
4) Another red square will appear.
5) Press on any of the squares, the read square must disappear.
6) Press again on any of the squares, the yellow square becomes black.

Then run without grab
  $ ./demo --without-grab

Once demo starts, a light-blue square appears. Then,
1) Press with a left mouse button on the square and release
2) Another yellow square will appear (it will be created as a popup window)
3) Press again on the appeared square or on the existing one.
4) Another red square will appear.
5) Press on any of the squares, the read square must disappear (but
   it does not disappear in mutter. Weston compositor reacts correctly, though).
6) Press again and mutter crashes (in Weston compositor, it becomes black).